update farming tab ui (#210)
