+++ WPFUI/Views/Tabs/FarmingPage.xaml
@@ -3,33 +3,42 @@
         x:TypeArguments="vms:FarmingViewModel"
         xmlns:rxui="http://reactiveui.net"
         xmlns:vms="clr-namespace:WPFUI.ViewModels.Tabs"
-        xmlns:uc="clr-namespace:WPFUI.Views.Uc"
-        xmlns:ucfarm="clr-namespace:WPFUI.Views.Uc.FarmingView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        mc:Ignorable="d"
-        d:DesignHeight="450" d:DesignWidth="800">
-
+        xmlns:controls="clr-namespace:WPFUI.Views.Uc">
     <Grid Background="#FFE5E5E5" Margin="0,2,0,0">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="1*" />
-            <ColumnDefinition Width="3*" />
-            <ColumnDefinition Width="1*" />
+            <ColumnDefinition Width="2*" />
+            <ColumnDefinition Width="2*" />
         </Grid.ColumnDefinitions>
         <Grid.RowDefinitions>
-            <RowDefinition />
+            <RowDefinition Height="1*" />
+            <RowDefinition Height="5*" />
+            <RowDefinition Height="5*" />
         </Grid.RowDefinitions>
-        <ucfarm:FarmListUc />
-        <ucfarm:FarmContentUc Grid.Column="1" />
-        <Grid Grid.Column="2" Grid.ColumnSpan="2">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="1*" />
-                <RowDefinition Height="1*" />
-            </Grid.RowDefinitions>
-            <Button x:Name="StartButton" Content="Start all active"  Margin="0,5,0,0" />
-            <Button x:Name="StopButton" Content="Stop all" Grid.Row="1" Margin="0,5,0,0" />
-        </Grid>
+
+        <!--Row 0-->
+        <Button x:Name="RefreshFarmListsButton" Content="Refresh farm lists" Margin="0,5,0,0" />
+        <controls:ToleranceUc x:Name="Interval" Grid.Column="1" />
+        <Button x:Name="SaveButton" Content="Save" Grid.Column="2" Margin="5,5,5,5" />
+        <!--Row 1-->
+        <Button x:Name="ActiveButton" Grid.Row="1" Grid.Column="1" Grid.ColumnSpan="2" Margin="5,5,5,5" />
+        <!--Row 2-->
+        <Button x:Name="StartButton" Content="Start all active" Grid.Row="2" Grid.Column="1" Margin="5,5,5,5" />
+        <Button x:Name="StopButton" Content="Stop all" Grid.Row="2" Grid.Column="2" Margin="5,5,5,5" />
+
+        <!--Side-->
+        <ListBox x:Name="FarmListViewer" SelectionMode="Single" IsSynchronizedWithCurrentItem="True" Grid.Row="1" Grid.RowSpan="2" Margin="0,5,0,0">
+            <ListBox.ItemTemplate>
+                <DataTemplate>
+                    <TextBlock Text="{Binding Content}">
+                        <TextBlock.Foreground>
+                            <SolidColorBrush Color="{Binding Color}" />
+                        </TextBlock.Foreground>
+                    </TextBlock>
+                </DataTemplate>
+            </ListBox.ItemTemplate>
+        </ListBox>
     </Grid>
 </rxui:ReactivePage>
