--- WPFUI/Views/Uc/FarmingView/FarmContentUc.xaml
@@ -1,22 +1,0 @@
-﻿<rxui:ReactiveUserControl
-        x:Class="WPFUI.Views.Uc.FarmingView.FarmContentUc"
-        x:TypeArguments="vms:FarmContentViewModel"
-        xmlns:rxui="http://reactiveui.net"
-        xmlns:vms="clr-namespace:WPFUI.ViewModels.Uc.FarmingView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:controls="clr-namespace:WPFUI.Views.Uc">
-    <Grid x:Name="MainGrid" Background="#FFE3E3E3">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="1*" />
-            <RowDefinition Height="1*" />
-            <RowDefinition Height="1*" />
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition />
-        </Grid.ColumnDefinitions>
-        <CheckBox x:Name="ActiveCheckBox" Content="Active this farm list" HorizontalAlignment="Center"  VerticalAlignment="Center" />
-        <controls:ToleranceUc x:Name="Interval" Grid.Row="1" />
-        <Button x:Name="SaveButton" Content="Save" Grid.Row="2" Margin="5,5,5,5" />
-    </Grid>
-</rxui:ReactiveUserControl>
