--- WPFUI/Views/Uc/FarmingView/FarmListUc.xaml
@@ -1,26 +1,0 @@
-﻿<rxui:ReactiveUserControl
-        x:Class="WPFUI.Views.Uc.FarmingView.FarmListUc"
-        x:TypeArguments="vms:FarmListViewModel"
-        xmlns:rxui="http://reactiveui.net"
-        xmlns:vms="clr-namespace:WPFUI.ViewModels.Uc.FarmingView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="1*" />
-            <RowDefinition Height="8*" />
-        </Grid.RowDefinitions>
-        <Button x:Name="RefreshFarmListsButton" Content="Refresh farm lists" Margin="0,5,0,0" />
-        <ListBox  x:Name="FarmListViewer" SelectionMode="Single" IsSynchronizedWithCurrentItem="True" Grid.Row="1" Margin="0,5,0,0">
-            <ListBox.ItemTemplate>
-                <DataTemplate>
-                    <TextBlock Text="{Binding Content}">
-                        <TextBlock.Foreground>
-                            <SolidColorBrush Color="{Binding Color}" />
-                        </TextBlock.Foreground>
-                    </TextBlock>
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>
-    </Grid>
-</rxui:ReactiveUserControl>
